refactor(components): move header/footer to layout folder
