+++ rspack-ui/src/pages/navPage.tsx
@@ -2,8 +2,8 @@ import React from 'react'
 
 import { Outlet } from 'react-router-dom';
 
-import TopHeader from '../components/Header';
-import Footer from '../components/Footer';
+import TopHeader from '../components/layout/Header';
+import Footer from '../components/layout/Footer';
 // This is the component that renders the header and footer for all pages
 export default function NavPage() {
     return (
